requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
aiohttp==3.9.1
websockets==12.0
python-dotenv==1.0.0
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0 